Allow using as a dependency in addition to running as a globally installed app
See README.md for an example.
