+++ Dockerfile
@@ -5,4 +5,4 @@ ADD . .
 RUN npm install
 
 EXPOSE 8001
-CMD ["node", "index.js"]
+CMD ["node", "bin/dynamodb-admin.js"]
