+++ package.json
@@ -2,7 +2,7 @@
   "name": "dynamodb-admin",
   "version": "2.0.4",
   "description": "GUI for DynamoDB. Useful for local development.",
-  "main": "index.js",
+  "main": "lib/backend.js",
   "bin": "./bin/dynamodb-admin.js",
   "scripts": {
     "start": "node bin/dynamodb-admin.js",
